refactor: Move get_folder_path function to utils.py
This commit organizes the code by moving the get_folder_path function from  the files models.py to a new utils.py file.

This improves code modularity and readability while allowing for easier reuse of the function in the future.
